+++ _tests/test_pmax.py
@@ -28,8 +28,8 @@ from boundary import (
 class pmax(unittest.TestCase):
 
   def test_01(self):
-    P0 = np.float64(15e6)
-    T0 = np.float64(68. + 273.15)
+    P0 = 15e6
+    T0 = 68. + 273.15
     yi = np.array([0.7167, 0.0895, 0.0917, 0.0448, 0.0573])
     Pci = np.array([45.99, 48.72, 42.48, 37.96, 23.975]) * 1e5
     Tci = np.array([190.56, 305.32, 369.83, 425.12, 551.022])
@@ -51,8 +51,8 @@ class pmax(unittest.TestCase):
     pass
 
   def test_02(self):
-    P0 = np.float64(20e6)
-    T0 = np.float64(68. + 273.15)
+    P0 = 20e6
+    T0 = 68. + 273.15
     yi = np.array([0.7167, 0.0895, 0.0917, 0.0448, 0.0573])
     Pci = np.array([45.99, 48.72, 42.48, 37.96, 23.975]) * 1e5
     Tci = np.array([190.56, 305.32, 369.83, 425.12, 551.022])
@@ -74,8 +74,8 @@ class pmax(unittest.TestCase):
     pass
 
   def test_03(self):
-    P0 = np.float64(15e6)
-    T0 = np.float64(68. + 273.15)
+    P0 = 15e6
+    T0 = 68. + 273.15
     yi = np.array([0.7167, 0.0895, 0.0917, 0.0448, 0.0573])
     Pci = np.array([45.99, 48.72, 42.48, 37.96, 23.975]) * 1e5
     Tci = np.array([190.56, 305.32, 369.83, 425.12, 551.022])
@@ -97,8 +97,8 @@ class pmax(unittest.TestCase):
     pass
 
   def test_04(self):
-    P0 = np.float64(20e6)
-    T0 = np.float64(68. + 273.15)
+    P0 = 20e6
+    T0 = 68. + 273.15
     yi = np.array([0.7167, 0.0895, 0.0917, 0.0448, 0.0573])
     Pci = np.array([45.99, 48.72, 42.48, 37.96, 23.975]) * 1e5
     Tci = np.array([190.56, 305.32, 369.83, 425.12, 551.022])
@@ -120,8 +120,8 @@ class pmax(unittest.TestCase):
     pass
 
   def test_05(self):
-    P0 = np.float64(15e6)
-    T0 = np.float64(68. + 273.15)
+    P0 = 15e6
+    T0 = 68. + 273.15
     yi = np.array([0.7167, 0.0895, 0.0917, 0.0448, 0.0573])
     Pci = np.array([45.99, 48.72, 42.48, 37.96, 23.975]) * 1e5
     Tci = np.array([190.56, 305.32, 369.83, 425.12, 551.022])
@@ -143,8 +143,8 @@ class pmax(unittest.TestCase):
     pass
 
   def test_06(self):
-    P0 = np.float64(20e6)
-    T0 = np.float64(68. + 273.15)
+    P0 = 20e6
+    T0 = 68. + 273.15
     yi = np.array([0.7167, 0.0895, 0.0917, 0.0448, 0.0573])
     Pci = np.array([45.99, 48.72, 42.48, 37.96, 23.975]) * 1e5
     Tci = np.array([190.56, 305.32, 369.83, 425.12, 551.022])
