+++ _tests/test_stabilityPT.py
@@ -29,8 +29,8 @@ from stability import (
 class stab(unittest.TestCase):
 
   def test_01(self):
-    P = np.float64(2e6)
-    T = np.float64(40. + 273.15)
+    P = 2e6
+    T = 40. + 273.15
     yi = np.array([.15, .85])
     Pci = np.array([7.37646e6, 4.600155e6])
     Tci = np.array([304.2, 190.6])
@@ -46,8 +46,8 @@ class stab(unittest.TestCase):
     pass
 
   def test_02(self):
-    P = np.float64(6e6)
-    T = np.float64(10. + 273.15)
+    P = 6e6
+    T = 10. + 273.15
     yi = np.array([.9, .1])
     Pci = np.array([7.37646e6, 4.600155e6])
     Tci = np.array([304.2, 190.6])
@@ -63,8 +63,8 @@ class stab(unittest.TestCase):
     pass
 
   def test_03(self):
-    P = np.float64(101325.)
-    T = np.float64(20. + 273.15)
+    P = 101325.
+    T = 20. + 273.15
     yi = np.array([.1, .6, .3])
     Pci = np.array([4.600155e6, 3.2890095e6, 22.04832e6])
     Tci = np.array([190.6, 507.5, 647.3])
@@ -80,8 +80,8 @@ class stab(unittest.TestCase):
     pass
 
   def test_04(self):
-    P = np.float64(17e6)
-    T = np.float64(68. + 273.15)
+    P = 17e6
+    T = 68. + 273.15
     yi = np.array([0.7167, 0.0895, 0.0917, 0.0448, 0.0573])
     Pci = np.array([4.599e6, 4.872e6, 4.248e6, 3.796e6, 2.398e6])
     Tci = np.array([190.56, 305.32, 369.83, 425.12, 551.02])
@@ -102,8 +102,8 @@ class stab(unittest.TestCase):
     pass
 
   def test_05(self):
-    P = np.float64(2e6)
-    T = np.float64(40. + 273.15)
+    P = 2e6
+    T = 40. + 273.15
     yi = np.array([.15, .85])
     Pci = np.array([7.37646e6, 4.600155e6])
     Tci = np.array([304.2, 190.6])
@@ -119,8 +119,8 @@ class stab(unittest.TestCase):
     pass
 
   def test_06(self):
-    P = np.float64(6e6)
-    T = np.float64(10. + 273.15)
+    P = 6e6
+    T = 10. + 273.15
     yi = np.array([.9, .1])
     Pci = np.array([7.37646e6, 4.600155e6])
     Tci = np.array([304.2, 190.6])
@@ -136,8 +136,8 @@ class stab(unittest.TestCase):
     pass
 
   def test_07(self):
-    P = np.float64(101325.)
-    T = np.float64(20. + 273.15)
+    P = 101325.
+    T = 20. + 273.15
     yi = np.array([.1, .6, .3])
     Pci = np.array([4.600155e6, 3.2890095e6, 22.04832e6])
     Tci = np.array([190.6, 507.5, 647.3])
@@ -153,8 +153,8 @@ class stab(unittest.TestCase):
     pass
 
   def test_08(self):
-    P = np.float64(17e6)
-    T = np.float64(68. + 273.15)
+    P = 17e6
+    T = 68. + 273.15
     yi = np.array([0.7167, 0.0895, 0.0917, 0.0448, 0.0573])
     Pci = np.array([4.599e6, 4.872e6, 4.248e6, 3.796e6, 2.398e6])
     Tci = np.array([190.56, 305.32, 369.83, 425.12, 551.02])
@@ -175,8 +175,8 @@ class stab(unittest.TestCase):
     pass
 
   def test_09(self):
-    P = np.float64(2e6)
-    T = np.float64(40. + 273.15)
+    P = 2e6
+    T = 40. + 273.15
     yi = np.array([.15, .85])
     Pci = np.array([7.37646e6, 4.600155e6])
     Tci = np.array([304.2, 190.6])
@@ -192,8 +192,8 @@ class stab(unittest.TestCase):
     pass
 
   def test_10(self):
-    P = np.float64(6e6)
-    T = np.float64(10. + 273.15)
+    P = 6e6
+    T = 10. + 273.15
     yi = np.array([.9, .1])
     Pci = np.array([7.37646e6, 4.600155e6])
     Tci = np.array([304.2, 190.6])
@@ -209,8 +209,8 @@ class stab(unittest.TestCase):
     pass
 
   def test_11(self):
-    P = np.float64(17e6)
-    T = np.float64(68. + 273.15)
+    P = 17e6
+    T = 68. + 273.15
     yi = np.array([0.7167, 0.0895, 0.0917, 0.0448, 0.0573])
     Pci = np.array([4.599e6, 4.872e6, 4.248e6, 3.796e6, 2.398e6])
     Tci = np.array([190.56, 305.32, 369.83, 425.12, 551.02])
@@ -231,8 +231,8 @@ class stab(unittest.TestCase):
     pass
 
   def test_12(self):
-    P = np.float64(17e6)
-    T = np.float64(68. + 273.15)
+    P = 17e6
+    T = 68. + 273.15
     yi = np.array([0.7167, 0.0895, 0.0917, 0.0448, 0.0573])
     Pci = np.array([4.599e6, 4.872e6, 4.248e6, 3.796e6, 2.398e6])
     Tci = np.array([190.56, 305.32, 369.83, 425.12, 551.02])
@@ -253,8 +253,8 @@ class stab(unittest.TestCase):
     pass
 
   def test_13(self):
-    P = np.float64(17e6)
-    T = np.float64(68. + 273.15)
+    P = 17e6
+    T = 68. + 273.15
     yi = np.array([0.7167, 0.0895, 0.0917, 0.0448, 0.0573])
     Pci = np.array([4.599e6, 4.872e6, 4.248e6, 3.796e6, 2.398e6])
     Tci = np.array([190.56, 305.32, 369.83, 425.12, 551.02])
@@ -275,8 +275,8 @@ class stab(unittest.TestCase):
     pass
 
   def test_14(self):
-    P = np.float64(1e6)
-    T = np.float64(160.)
+    P = 1e6
+    T = 160.
     yi = np.array([0.9430, 0.0270, 0.0074, 0.0049, 0.0027, 0.0010, 0.0140])
     Pci = np.array([4.599, 4.872, 4.248, 3.796, 3.370, 3.025, 3.400]) * 1e6
     Tci = np.array([190.56, 305.32, 369.83, 425.12, 469.70, 507.60, 126.20])
@@ -299,8 +299,8 @@ class stab(unittest.TestCase):
     pass
 
   def test_15(self):
-    P = np.float64(1e6)
-    T = np.float64(160.)
+    P = 1e6
+    T = 160.
     yi = np.array([0.9430, 0.0270, 0.0074, 0.0049, 0.0027, 0.0010, 0.0140])
     Pci = np.array([4.599, 4.872, 4.248, 3.796, 3.370, 3.025, 3.400]) * 1e6
     Tci = np.array([190.56, 305.32, 369.83, 425.12, 469.70, 507.60, 126.20])
@@ -323,8 +323,8 @@ class stab(unittest.TestCase):
     pass
 
   def test_16(self):
-    P = np.float64(1e6)
-    T = np.float64(160.)
+    P = 1e6
+    T = 160.
     yi = np.array([0.9430, 0.0270, 0.0074, 0.0049, 0.0027, 0.0010, 0.0140])
     Pci = np.array([4.599, 4.872, 4.248, 3.796, 3.370, 3.025, 3.400]) * 1e6
     Tci = np.array([190.56, 305.32, 369.83, 425.12, 469.70, 507.60, 126.20])
@@ -347,8 +347,8 @@ class stab(unittest.TestCase):
     pass
 
   def test_17(self):
-    P = np.float64(8.7625e6)
-    T = np.float64(316.48)
+    P = 8.7625e6
+    T = 316.48
     yi = np.array([
       0.590516, 0.028933, 0.072729, 0.081162, 0.131012, 0.064671, 0.030979,
     ])
@@ -373,8 +373,8 @@ class stab(unittest.TestCase):
     pass
 
   def test_18(self):
-    P = np.float64(8.7625e6)
-    T = np.float64(316.48)
+    P = 8.7625e6
+    T = 316.48
     yi = np.array([
       0.590516, 0.028933, 0.072729, 0.081162, 0.131012, 0.064671, 0.030979,
     ])
@@ -399,8 +399,8 @@ class stab(unittest.TestCase):
     pass
 
   def test_19(self):
-    P = np.float64(8.7625e6)
-    T = np.float64(316.48)
+    P = 8.7625e6
+    T = 316.48
     yi = np.array([
       0.590516, 0.028933, 0.072729, 0.081162, 0.131012, 0.064671, 0.030979,
     ])
@@ -425,8 +425,8 @@ class stab(unittest.TestCase):
     pass
 
   def test_20(self):
-    P = np.float64(8.7625e6)
-    T = np.float64(316.48)
+    P = 8.7625e6
+    T = 316.48
     yi = np.array([
       0.590516, 0.028933, 0.072729, 0.081162, 0.131012, 0.064671, 0.030979,
     ])
@@ -451,8 +451,8 @@ class stab(unittest.TestCase):
     pass
 
   def test_21(self):
-    P = np.float64(8.7625e6)
-    T = np.float64(316.48)
+    P = 8.7625e6
+    T = 316.48
     yi = np.array([
       0.590516, 0.028933, 0.072729, 0.081162, 0.131012, 0.064671, 0.030979,
     ])
@@ -501,8 +501,8 @@ class stab(unittest.TestCase):
     pass
 
   def test_23(self):
-    P = np.float64(16.5e6)
-    T = np.float64(68. + 273.15)
+    P = 16.5e6
+    T = 68. + 273.15
     yi = np.array([0.7167, 0.0895, 0.0917, 0.0448, 0.0573])
     Pci = np.array([45.99, 48.72, 42.48, 37.96, 23.975]) * 1e5
     Tci = np.array([190.56, 305.32, 369.83, 425.12, 551.022])
@@ -523,8 +523,8 @@ class stab(unittest.TestCase):
     pass
 
   def test_24(self):
-    P = np.float64(16e6)
-    T = np.float64(62.68 + 273.15)
+    P = 16e6
+    T = 62.68 + 273.15
     yi = np.array([0.7167, 0.0895, 0.0917, 0.0448, 0.0573])
     Pci = np.array([45.99, 48.72, 42.48, 37.96, 23.975]) * 1e5
     Tci = np.array([190.56, 305.32, 369.83, 425.12, 551.022])
